add review update function, review delete function, fix note to float system

+++ app/Http/Requests/ReviewRequest.php
@@ -17,7 +17,7 @@ class ReviewRequest extends FormRequest
     {
         return [
             'comment' => ['required', 'string', 'max:255', 'min:3'],
-            'note' => ['required', 'integer', 'min:1', 'max:5'],
+            'note' => ['required', 'numeric', 'min:1', 'max:5'],
             'movie_id' => ['required'],
         ];
     }
